:bug: feat: fixed Sign in with Google
