+++ app/src/main/java/com/example/mobdev21_night_at_the_museum/MainActivity.java
@@ -2,50 +2,16 @@ package com.example.mobdev21_night_at_the_museum;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.mobdev21_night_at_the_museum.R;
-import com.example.mobdev21_night_at_the_museum.presentation.intro.LoginActivity;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
+import com.example.mobdev21_night_at_the_museum.presentation.intro.IntroductionActivity;
 
 public class MainActivity extends AppCompatActivity {
-    Button logout;
-    FirebaseUser user;
-    private FirebaseAuth mAuth;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        mAuth = FirebaseAuth.getInstance();
-        user = mAuth.getCurrentUser();
-        if (user == null) {
-            Intent auth = new Intent(getApplicationContext(), LoginActivity.class);
-            startActivity(auth);
-            finish();
-        }
-        logout = findViewById(R.id.signout);
-
-        logout.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                signOut();
-            }
-        });
-
+        Intent introductionScreen = new Intent(getApplicationContext(), IntroductionActivity.class);
+        startActivity(introductionScreen);
     }
-
-    private void signOut() {
-
-        FirebaseAuth.getInstance().signOut();
-
-        Intent toLogin = new Intent(getApplicationContext(), LoginActivity.class);
-        startActivity(toLogin);
-        finish();
-    }
-
 }
